requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.4
sqlalchemy==2.0.25
openai==1.14.3
python-dotenv==1.0.1
plotly==5.19.0
matplotlib==3.8.2
seaborn==0.13.2
openpyxl==3.1.2